Reorganise classes into packages - command, managers, misc, task
