--- src/main/java/ByeCommand.java
@@ -1,16 +1,0 @@
-public class ByeCommand extends Command{
-    int taskIdx;
-    FileManager fileManager;
-
-    public ByeCommand(String userTaskString) {
-        super(userTaskString);
-        this.fileManager = fileManager;
-    }
-
-    public void executeTask(TaskList taskList, FileManager fileManager) throws DukeException {
-        fileManager.saveTasks();
-    }
-
-}
-
-
