+++ src/main/java/duke/Ui.java
@@ -1,14 +1,27 @@
+package duke;
+
 public class Ui {
 
+    /**
+     * Print statement with divider line
+     * @param str
+     */
     public void print(Object str) {
         System.out.println(str);
         System.out.println("_".repeat(100));
     }
 
+    /**
+     * Prints oridinary System.out.println
+     * @param str
+     */
     public void println(Object str) {
         System.out.println(str);
     }
 
+    /**
+     * Print statemeny for when system exits
+     */
     public void printBye(){
         this.print("Bye. Hope I've motivated you as much as I could have, and SMILE :D");
     }
