+++ src/main/java/duke/command/AddCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.managers.FileManager;
+import duke.managers.TaskList;
+import duke.task.Task;
+
 public class AddCommand extends Command {
     Task task;
 
@@ -6,6 +12,11 @@ public class AddCommand extends Command {
         this.task = task;
     }
 
+    /**
+     * Executes adding of task to TaskList (with printing)
+     * @param taskList
+     * @param fileManager
+     */
     public void executeTask(TaskList taskList, FileManager fileManager) {
         taskList.addTask(this.task, true);
     }
