+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.managers.FileManager;
+import duke.managers.TaskList;
+
 public class DeleteCommand extends Command{
     int taskIdx;
 
@@ -6,6 +12,12 @@ public class DeleteCommand extends Command{
         this.taskIdx = taskIdx;
     }
 
+    /**
+     * Executes deletion of task per taskIdx
+     * @param taskList
+     * @param fileManager
+     * @throws DukeException if task idx is not valid
+     */
     public void executeTask(TaskList taskList, FileManager fileManager) throws DukeException {
         taskList.deleteTask(this.taskIdx);
     }
