+++ src/main/java/duke/command/ListCommand.java
@@ -1,3 +1,8 @@
+package duke.command;
+
+import duke.managers.FileManager;
+import duke.managers.TaskList;
+
 public class ListCommand extends Command{
     int taskIdx;
 
@@ -5,6 +10,11 @@ public class ListCommand extends Command{
         super(userTaskString);
     }
 
+    /**
+     * List current remaining user tasks.
+     * @param taskList
+     * @param fileManager
+     */
     public void executeTask(TaskList taskList, FileManager fileManager) {
         taskList.printUserTasks();
     }
