+++ src/main/java/duke/command/MarkCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.managers.FileManager;
+import duke.managers.TaskList;
+
 public class MarkCommand extends Command{
     int taskIdx;
 
@@ -6,6 +12,12 @@ public class MarkCommand extends Command{
         this.taskIdx = taskIdx;
     }
 
+    /**
+     * Marks user task as per taskIdx.
+     * @param taskList
+     * @param fileManager
+     * @throws DukeException
+     */
     public void executeTask(TaskList taskList, FileManager fileManager) throws DukeException {
         taskList.markTaskDone(taskIdx, true);
     }
