+++ src/main/java/duke/task/Task.java
@@ -1,3 +1,4 @@
+package duke.task;
 public abstract class Task {
     private boolean isDone;
     private String taskDescription;
